requirements.txt
argparse==1.3.0
Django==1.8.3
djangorestframework==3.3.3
wheel==0.24.0
